Encode-DecodeWithString
Here is a code snippet to help you encrypt and decrypt a string in C#. The key is hardcoded, but in your own project you may want to read it from a secure store, or perhaps a config file or environment variable. Notice that the key is being hashed for an extra layer of security.